+++ src/parser/enrichers/item/Candle.mjs
@@ -1,13 +1,18 @@
 /* eslint-disable class-methods-use-this */
 import DDBEnricherData from "../data/DDBEnricherData.mjs";
+import GenericLightSource from "./GenericLightSource.mjs";
 
-export default class Candle extends DDBEnricherData {
+export default class Candle extends GenericLightSource {
 
   get effects() {
     return [
       {
         atlOnly: true,
         name: "Candle Light",
+        activityMatch: "Light",
+        options: {
+          transfer: false,
+        },
         atlChanges: [
           DDBEnricherData.ChangeHelper.atlChange("ATL.light.dim", CONST.ACTIVE_EFFECT_MODES.UPGRADE, '5'),
           DDBEnricherData.ChangeHelper.atlChange("ATL.light.bright", CONST.ACTIVE_EFFECT_MODES.UPGRADE, '10'),
@@ -16,6 +21,16 @@ export default class Candle extends DDBEnricherData {
         ],
       },
     ];
+  }
+
+  get override() {
+    return {
+      data: {
+        "system.uses": {
+          autoDestroy: true,
+        },
+      },
+    };
   }
 
 }