+++ src/parser/enrichers/item/GenericLightSource.mjs
@@ -1,0 +1,13 @@
+/* eslint-disable class-methods-use-this */
+import DDBEnricherData from "../data/DDBEnricherData.mjs";
+
+export default class GenericLightSource extends DDBEnricherData {
+
+  get activity() {
+    return {
+      name: "Light",
+      noTemplate: true,
+    };
+  }
+
+}