+++ src/parser/enrichers/item/Torch.mjs
@@ -1,7 +1,8 @@
 /* eslint-disable class-methods-use-this */
 import DDBEnricherData from "../data/DDBEnricherData.mjs";
+import GenericLightSource from "./GenericLightSource.mjs";
 
-export default class Torch extends DDBEnricherData {
+export default class Torch extends GenericLightSource {
 
   get effects() {
     const lightAnimation = '{type: "torch", speed: 2, intensity: 2}';
@@ -9,6 +10,11 @@ export default class Torch extends DDBEnricherData {
       {
         atlOnly: true,
         name: "Torch Light",
+        activityMatch: "Light",
+        options: {
+          transfer: false,
+          durationSeconds: 3600,
+        },
         atlChanges: [
           DDBEnricherData.ChangeHelper.atlChange("ATL.light.dim", CONST.ACTIVE_EFFECT_MODES.UPGRADE, '40'),
           DDBEnricherData.ChangeHelper.atlChange("ATL.light.bright", CONST.ACTIVE_EFFECT_MODES.UPGRADE, '20'),
@@ -19,5 +25,16 @@ export default class Torch extends DDBEnricherData {
       },
     ];
   }
+
+  get override() {
+    return {
+      data: {
+        "system.uses": {
+          autoDestroy: true,
+        },
+      },
+    };
+  }
+
 
 }