+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Water Bullet spell fixes.
 - Mantle of Majesty no longer parsed correctly. @_mickael
+- Improve ATL/ATE effect for Torch and other consumable items that cast light.
 
 # 6.4.18